refactor(sourcepoint_unified_cmp_platform_interface): move interface implementation to separate file for better code organization feat(sourcepoint_unified_cmp_platform_interface): add new method channel implementation for better platform communication refactor(sourcepoint_unified_cmp_platform_interface): enhance types documentation for better code readability and understanding
